Adjust entities (id type Integer->Long; id->nameId)

+++ src/main/java/pl/polsl/io/charityapp/repository/CharityActionRepository.java
@@ -6,7 +6,7 @@ import pl.polsl.io.charityapp.model.entity.CharityAction;
 import java.util.List;
 import java.util.Optional;
 
-public interface CharityActionRepository extends JpaRepository<CharityAction, Integer> {
+public interface CharityActionRepository extends JpaRepository<CharityAction, Long> {
     List<CharityAction> findAll();
 
     Optional<CharityAction> findByName(String name);
